feat: add input layer expansion
